--- src/main/java/no/nav/fo/veilarbdialog/oppfolging/v2/OppfolgingV2UnderOppfolgingDTO.java
@@ -1,8 +1,0 @@
-package no.nav.fo.veilarbdialog.oppfolging.v2;
-
-import lombok.Value;
-
-@Value
-public class OppfolgingV2UnderOppfolgingDTO {
-    boolean erUnderOppfolging;
-}
